Revert "[CodeGen] Introduce `MachineDomTreeUpdater` (#95369)"
This reverts commit 6ca387cbcb207abe2a07bbb1b536f099c2e246e7.
