+++ llvm/lib/CodeGen/MachineBasicBlock.cpp
@@ -315,12 +315,6 @@ bool MachineBasicBlock::isLegalToHoistInto() const {
   return true;
 }
 
-bool MachineBasicBlock::hasName() const {
-  if (const BasicBlock *LBB = getBasicBlock())
-    return LBB->hasName();
-  return false;
-}
-
 StringRef MachineBasicBlock::getName() const {
   if (const BasicBlock *LBB = getBasicBlock())
     return LBB->getName();
